Fix workspace deletion edge cases
