+++ include/sway/tree/layout.h
@@ -39,6 +39,8 @@ struct sway_container *container_add_sibling(struct sway_container *parent,
 
 struct sway_container *container_remove_child(struct sway_container *child);
 
+struct sway_container *container_reap_empty(struct sway_container *container);
+
 void container_move_to(struct sway_container* container,
 		struct sway_container* destination);
