+++ include/sway/tree/workspace.h
@@ -23,4 +23,6 @@ struct sway_container *workspace_output_prev(struct sway_container *current);
 
 struct sway_container *workspace_prev(struct sway_container *current);
 
+bool workspace_is_visible(struct sway_container *ws);
+
 #endif
