+++ sway/meson.build
@@ -82,6 +82,7 @@ sway_sources = files(
 	'security.c',
 	'tree/container.c',
 	'tree/layout.c',
+	'tree/output.c',
 	'tree/view.c',
 	'tree/workspace.c',
 )
